added soft delete

+++ lib/removeDoc.ts
@@ -9,7 +9,15 @@ export default async function RemoveDoc({ id }: { id: number }) {
   const con = connect(config);
   const db = drizzle(con);
 
-  const doc = await db.delete(documents).where(eq(documents.id, id));
+  const doc = await db.select().from(documents).where(eq(documents.id, id));
 
-  return doc;
+  if (doc[0].deleted === 1) {
+    // Soft Delete
+    return await db.delete(documents).where(eq(documents.id, id));
+  } else {
+    // Hard Delete
+    await db.update(documents).set({ deleted: 1 }).where(eq(documents.id, id));
+  }
+
+  return {};
 }
